CIT-9 stat: fixed mapping error

+++ src/main/java/Citronix/dto/records/farm/FarmRequestDTO.java
@@ -17,6 +17,6 @@ public record FarmRequestDTO(
 
         @NotNull
         @Min(value = 1, message = "Value must be >= 1")
-        long superficie
+        double superficie
 ) {
 }
