+++ src/main/java/Citronix/dto/records/farm/FarmResponseDTO.java
@@ -20,6 +20,6 @@ public record FarmResponseDTO(
 
         @NotNull
         @Min(value = 1, message = "Value must be >= 1")
-        long superficie
+        double superficie
 ) {
 }
