+++ src/main/java/Citronix/dto/records/field/FieldUpdateDTO.java
@@ -4,6 +4,6 @@ import jakarta.validation.constraints.NotNull;
 
 public record FieldUpdateDTO(
         @NotNull
-        long superficie
+        double superficie
 ) {
 }
